Shipment now creates an item in the inventory
Shipment now uses PackageData to store information about package, and now attempts to creates an item in the inventory based on the data it has.
Duplicated and edited CreateItem, one is now CreateRandomItem (no input variables) and the other is CreateItem (PackageData input)
Duplicated and edited InsertItem. New version automatically creates the item to that is inserted